+++ src/electron.h
@@ -17,7 +17,7 @@ typedef struct {
 electron_t* electron_create(int posx, int posy, float radius);
 void electron_destroy(electron_t * electron);
 void electron_render(electron_t * electron, SDL_Renderer * renderer, SDL_Color * color);
-void electron_move(electron_t * electron, int amountx, int amounty);
+void electron_move(electron_t * electron, float amountx, float amounty);
 void electron_apply_force(electron_t *electron, force_t *force, float duration);
 void electron_update(electron_t* electron, float dt);
 #endif
